% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosketch.R
\name{geosketch}
\alias{geosketch}
\title{Run geosketch to subsample a matrix}
\usage{
geosketch(mat, N, replace = FALSE, one_indexed = TRUE)
}
\arguments{
\item{mat}{m x n matrix. Samples (the dimension along which to subsample)
should be in the rows, features in the columns.}

\item{N}{Numeric scalar, the number of samples to retain.}

\item{replace}{Logical scalar, whether to sample with replacement.}

\item{one_indexed}{Logical scalar, whether to return one-indexed indices.}
}
\value{
A numeric vector with indices to retain.
}
\description{
Run geosketch to subsample a matrix
}
\author{
Charlotte Soneson
}
