% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosketch.R
\name{getGeosketchNames}
\alias{getGeosketchNames}
\title{Get names of geosketch functions}
\usage{
getGeosketchNames()
}
\value{
A list of names of objects exposed in the geosketch module
}
\description{
Get names of geosketch functions
}
\examples{
getGeosketchNames()

}
\author{
Charlotte Soneson
}
