% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsampler.R
\name{getScSamplerNames}
\alias{getScSamplerNames}
\title{Get names of scSampler functions}
\usage{
getScSamplerNames()
}
\value{
A list of names of objects exposed in the scSampler module
}
\description{
Get names of scSampler functions
}
\examples{
if (!(Sys.info()["sysname"] == "Darwin" && Sys.info()["machine"] == "arm64")) {
    getScSamplerNames()
}

}
\author{
Charlotte Soneson
}
