% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchR-package.R
\docType{package}
\name{sketchR-package}
\alias{sketchR}
\alias{sketchR-package}
\title{sketchR - an R interface for python subsampling/sketching algorithms}
\description{
The sketchR package provides an R interface for various subsampling
algorithms implemented in python packages. Currently, interfaces to the
geosketch and scSampler python packages are implemented, in the functions
geosketch() and scsampler(), respectively. In addition the package also
provides diagnostic plots to evaluate the subsampling.
More details on how to get started and incorporate the subsampling into
an scRNA-seq workflow are provided in the vignette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmicompbio/sketchR}
  \item Report bugs at \url{https://github.com/fmicompbio/sketchR/issues}
}

}
\author{
Charlotte Soneson

Michael Stadler
}
\keyword{internal}
