% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausdorffDistPlot.R
\name{hausdorffDistPlot}
\alias{hausdorffDistPlot}
\title{Create diagnostic plot of Hausdorff distances}
\usage{
hausdorffDistPlot(mat, Nvec, Nrep = 5, q = 1e-04, doPlot = TRUE, ...)
}
\arguments{
\item{mat}{m x n matrix. Samples (the dimension along which to subsample)
should be in the rows, features in the columns.}

\item{Nvec}{Numeric vector of sketch sizes.}

\item{Nrep}{Numeric scalar indicating the number of sketches to draw 
for each sketch size.}

\item{q}{Numeric scalar in [0,1], indicating the fraction of largest 
minimum distances to discard when calculating the robust Hausdorff 
distance. Setting q=0 gives the classical Hausdorff distance. 
The default is 1e-4, as suggested by Hie et al (2019).}

\item{doPlot}{Logical scalar, indicating whether or not to generate the 
plot.}

\item{...}{Additional arguments provided to \code{geosketch}.}
}
\value{
A \code{data.frame} with three columns: the size of the sketch
as the number of samples (N), the size of the sketch as a fraction 
of the original data set (frac), and the robust Hausdorff distance
(HausdorffDist).
}
\description{
Create diagnostic plot showing the Hausdorff distance between a sketch 
and the full data set, for varying sketch sizes.
}
\examples{
mat <- matrix(rnorm(1000), nrow = 100)
hdp <- hausdorffDistPlot(mat, Nvec = c(10, 25, 50))

}
\references{
Hie et al (2019): Geometric sketching compactly summarizes the
single-cell transcriptomic landscape. Cell Systems 8, 483–493.

Huttenlocher et al (1993): Comparing images using the Hausdorff
distance. IEEE Transactions on Pattern Analysis and Machine 
Intelligence 15(9), 850-863.
}
\author{
Charlotte Soneson
}
